% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climate-scraper.R
\name{get_station_data}
\alias{get_station_data}
\title{Data scraper for data from active stations in the Scripps Project}
\usage{
get_station_data(stationURL, type = "daily_flask_co2",
  dataURL = "http://scrippsco2.ucsd.edu")
}
\arguments{
\item{baseURL}{url for station information}
}
\description{
Information for active stations is available at http://scrippsco2.ucsd.edu/data/atmospheric_co2/sampling_stations
}
\examples{
safe_load <- purrr::safely(get_station_data)

library(wilydata)
scripps_data_co2 <- scripps_stations$Page \%>\%
  purrr::map_df(.f = function(s) {
   res <- safe_load(s, type="daily_flask_co2")
   if (!is.null(res$result)) {
     res$result$station_id <- basename(s)
   }
   res$result
 })

save(scripps_data_co2, file="../wilydata/data/scripps_data_co2.rda")
}
