% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noaa-weather-scraper.R
\name{noaa_api_data}
\alias{noaa_api_data}
\title{Get data from NOAA weather API for Ames, IA}
\usage{
noaa_api_data(location = NULL, var = c("TMAX", "TMIN"),
  date_min = "2008-01-01", date_max = Sys.Date(), quiet = T)
}
\arguments{
\item{var}{variables to download (TMAX, TMIN by default, see rnoaa 
documentation for other options)}

\item{date_min}{ymd format, minimum date}

\item{date_max}{ymd format, maximum date}

\item{quiet}{print messages?}

\item{lat_long_df}{data frame with columns id, latitude, and longitude}
}
\description{
Get data from NOAA weather API for Ames, IA
}
