% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climate-scraper.R
\name{scrape_scripps_stations}
\alias{scrape_scripps_stations}
\title{Data scraper for active stations in the Scripps Project}
\usage{

  scrape_scripps_stations(baseURL = "http://scrippsco2.ucsd.edu/data/atmospheric_co2")
}
\arguments{
\item{baseURL}{url for station information}
}
\description{
Information for active stations is available at http://scrippsco2.ucsd.edu/data/atmospheric_co2/sampling_stations
}
\examples{
scripps_stations <- scrape_scripps_stations()
# save station info into file
save(scripps_stations, file="../wilydata/data/scripps_stations.rda")
}
