% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wunderground-weather-scraper.R
\name{wUnderground_data}
\alias{wUnderground_data}
\title{Function to get all weather data from wunderground for a period of time}
\usage{
wUnderground_data(start = lubridate::ymd("2011-01-01"),
  end = lubridate::today(), station = "KAMW")
}
\arguments{
\item{start}{start date}

\item{end}{end date}

\item{station}{weather station}

\item{file}{file to save the CSV data file}
}
\description{
Function to get all weather data from wunderground for a period of time
}
