% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wunderground-weather-scraper.R
\name{wUnderground_day}
\alias{wUnderground_day}
\title{Function to get weather data from weather underground}
\usage{
wUnderground_day(remDr, station = "KAMW", year = 2015, month = 1,
  day = 1)
}
\arguments{
\item{remDr}{a selenium webdriver client}

\item{station}{station call sign}

\item{year}{year (4 digit number)}

\item{month}{month (1 or 2 digit number)}

\item{day}{day (1 or 2 digit number)}
}
\value{
table of weather data, with date, time, variable, value, and unit
}
\description{
Function to get weather data from weather underground
}
